robot_voice_control/requirements.txt
openai-agents[voice]
numpy
sounddevice
websocket-client
pyaudio